add partner badge to all data set list titles
