+++ src/components/molecules/PoolTransactions.tsx
@@ -4,7 +4,7 @@ import React, { ReactElement, useEffect, useState } from 'react'
 import EtherscanLink from '../atoms/EtherscanLink'
 import Time from '../atoms/Time'
 import Table from '../atoms/Table'
-import AssetTitle from './AssetTitle'
+import AssetTitle from './AssetListTitle'
 import styles from './PoolTransactions.module.css'
 import { formatCurrency } from '@coingecko/cryptoformat'
 import { useUserPreferences } from '../../providers/UserPreferences'
